LayoutTests/storage/indexeddb/modern/basic-put-expected.txt

ALERT: Upgrade needed: Old version - 0 New version - 1
ALERT: [object IDBTransaction] - versionchange
ALERT: [object IDBDatabase]
ALERT: put succeeded - key was 'foo'
ALERT: get succeeded - value was 'bar'
ALERT: version change transaction completed
PASS successfullyParsed is true

TEST COMPLETE
